+++ NEWS.md
@@ -1,3 +1,5 @@
+# eriplots (development version)
+
 # eriplots 0.1.2
 
 ## Added